requirements.txt
albumentations==0.3.3
jupyterlab==1.0.1
numpy==1.17.4
opencv-python==4.1.1.26
pandas==0.25.3
Pillow==6.2.1
plotly==4.3.0
pytorch-toolbelt==0.2.0
scikit-image==0.15.0
scikit-learn==0.21.3
scikit-multilearn==0.2.0
scipy==1.3.3
seaborn==0.9.0
segmentation-models-pytorch==0.0.3
skorch==0.6.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
torch==1.3.0
torchvision==0.4.0
tqdm==4.37.0
warmup-scheduler==0.1
xgboost==0.81
